requirements.txt
fastapi
uvicorn
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
httpx
# Force rebuild v1